--- classes/student.rb
@@ -1,9 +1,0 @@
-class Student < Person
-  def initialize(classroom)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    '¯(ツ)/¯'
-  end
-end
