--- classes/teacher.rb
@@ -1,5 +1,0 @@
-class Teacher < Person
-  def initialize(specialization)
-    @specialization = specialization
-  end
-end
