+++ person.rb
@@ -1,0 +1,22 @@
+class Person
+  attr_accessor :id, :name, :age
+
+  def initialize(age, name: 'Unknown', parent_permission: true)
+    @id = id
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+  end
+
+  private
+
+  def of_age?
+    @age >= 18
+  end
+
+  public
+
+  def can_use_services?
+    is_of_age? || @parent_permission
+  end
+end
